r/aws_cloudformation_stack: Start to tidy up waiters.
