+++ internal/service/cloudformation/wait.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -98,167 +97,6 @@ func WaitStackSetOperationSucceeded(ctx context.Context, conn *cloudformation.Cl
 }
 
 const (
-	// Default maximum amount of time to wait for a Stack to be Created
-	StackCreatedDefaultTimeout = 30 * time.Minute
-
-	stackCreatedMinTimeout = 1 * time.Second
-
-	// Default maximum amount of time to wait for a Stack to be Updated
-	StackUpdatedDefaultTimeout = 30 * time.Minute
-
-	stackUpdatedMinTimeout = 5 * time.Second
-
-	// Default maximum amount of time to wait for a Stack to be Deleted
-	StackDeletedDefaultTimeout = 30 * time.Minute
-
-	stackDeletedMinTimeout = 5 * time.Second
-)
-
-func WaitStackCreated(ctx context.Context, conn *cloudformation.CloudFormation, stackID, requestToken string, timeout time.Duration) (*cloudformation.Stack, error) {
-	stateConf := retry.StateChangeConf{
-		Pending: []string{
-			cloudformation.StackStatusCreateInProgress,
-			cloudformation.StackStatusDeleteInProgress,
-			cloudformation.StackStatusRollbackInProgress,
-		},
-		Target: []string{
-			cloudformation.StackStatusCreateComplete,
-			cloudformation.StackStatusCreateFailed,
-			cloudformation.StackStatusDeleteComplete,
-			cloudformation.StackStatusDeleteFailed,
-			cloudformation.StackStatusRollbackComplete,
-			cloudformation.StackStatusRollbackFailed,
-		},
-		Timeout:    timeout,
-		MinTimeout: stackCreatedMinTimeout,
-		Delay:      10 * time.Second,
-		Refresh:    StatusStack(ctx, conn, stackID),
-	}
-
-	outputRaw, err := stateConf.WaitForStateContext(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	stack, ok := outputRaw.(*cloudformation.Stack)
-	if !ok {
-		return nil, err
-	}
-
-	lastStatus := aws.StringValue(stack.StackStatus)
-	switch lastStatus {
-	// This will be the case if either disable_rollback is false or on_failure is ROLLBACK
-	case cloudformation.StackStatusRollbackComplete, cloudformation.StackStatusRollbackFailed:
-		reasons, err := getRollbackReasons(ctx, conn, stackID, requestToken)
-		if err != nil {
-			return stack, fmt.Errorf("failed to create CloudFormation stack, rollback requested (%s). Got an error reading failure information: %w", lastStatus, err)
-		}
-		return stack, fmt.Errorf("failed to create CloudFormation stack, rollback requested (%s): %q", lastStatus, reasons)
-
-	// This will be the case if on_failure is DELETE
-	case cloudformation.StackStatusDeleteComplete, cloudformation.StackStatusDeleteFailed:
-		reasons, err := getDeletionReasons(ctx, conn, stackID, requestToken)
-		if err != nil {
-			return stack, fmt.Errorf("failed to create CloudFormation stack, delete requested (%s). Got an error reading failure information: %w", lastStatus, err)
-		}
-
-		return stack, fmt.Errorf("failed to create CloudFormation stack, delete requested (%s): %q", lastStatus, reasons)
-
-	// This will be the case if either disable_rollback is true or on_failure is DO_NOTHING
-	case cloudformation.StackStatusCreateFailed:
-		reasons, err := getFailures(ctx, conn, stackID, requestToken)
-		if err != nil {
-			return stack, fmt.Errorf("failed to create CloudFormation stack (%s). Got an error reading failure information: %w", lastStatus, err)
-		}
-		return stack, fmt.Errorf("failed to create CloudFormation stack (%s): %q", lastStatus, reasons)
-	}
-
-	return stack, nil
-}
-
-func WaitStackUpdated(ctx context.Context, conn *cloudformation.CloudFormation, stackID, requestToken string, timeout time.Duration) (*cloudformation.Stack, error) {
-	stateConf := retry.StateChangeConf{
-		Pending: []string{
-			cloudformation.StackStatusUpdateCompleteCleanupInProgress,
-			cloudformation.StackStatusUpdateInProgress,
-			cloudformation.StackStatusUpdateRollbackInProgress,
-			cloudformation.StackStatusUpdateRollbackCompleteCleanupInProgress,
-		},
-		Target: []string{
-			cloudformation.StackStatusCreateComplete,
-			cloudformation.StackStatusUpdateComplete,
-			cloudformation.StackStatusUpdateRollbackComplete,
-			cloudformation.StackStatusUpdateRollbackFailed,
-		},
-		Timeout:    timeout,
-		MinTimeout: stackUpdatedMinTimeout,
-		Delay:      10 * time.Second,
-		Refresh:    StatusStack(ctx, conn, stackID),
-	}
-
-	outputRaw, err := stateConf.WaitForStateContext(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	stack, ok := outputRaw.(*cloudformation.Stack)
-	if !ok {
-		return nil, err
-	}
-
-	lastStatus := aws.StringValue(stack.StackStatus)
-	if lastStatus == cloudformation.StackStatusUpdateRollbackComplete || lastStatus == cloudformation.StackStatusUpdateRollbackFailed {
-		reasons, err := getRollbackReasons(ctx, conn, stackID, requestToken)
-		if err != nil {
-			return stack, fmt.Errorf("failed to update CloudFormation stack (%s). Got an error reading failure information: %w", lastStatus, err)
-		}
-
-		return stack, fmt.Errorf("failed to update CloudFormation stack (%s): %q", lastStatus, reasons)
-	}
-
-	return stack, nil
-}
-
-func WaitStackDeleted(ctx context.Context, conn *cloudformation.CloudFormation, stackID, requestToken string, timeout time.Duration) (*cloudformation.Stack, error) {
-	stateConf := retry.StateChangeConf{
-		Pending: []string{
-			cloudformation.StackStatusDeleteInProgress,
-			cloudformation.StackStatusRollbackInProgress,
-		},
-		Target: []string{
-			cloudformation.StackStatusDeleteComplete,
-			cloudformation.StackStatusDeleteFailed,
-		},
-		Timeout:    timeout,
-		MinTimeout: stackDeletedMinTimeout,
-		Delay:      10 * time.Second,
-		Refresh:    StatusStack(ctx, conn, stackID),
-	}
-
-	outputRaw, err := stateConf.WaitForStateContext(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	stack, ok := outputRaw.(*cloudformation.Stack)
-	if !ok {
-		return nil, err
-	}
-
-	lastStatus := aws.StringValue(stack.StackStatus)
-	if lastStatus == cloudformation.StackStatusDeleteFailed {
-		reasons, err := getFailures(ctx, conn, stackID, requestToken)
-		if err != nil {
-			return stack, fmt.Errorf("failed to delete CloudFormation stack (%s). Got an error reading failure information: %w", lastStatus, err)
-		}
-
-		return stack, fmt.Errorf("failed to delete CloudFormation stack (%s): %q", lastStatus, reasons)
-	}
-
-	return stack, nil
-}
-
-const (
 	TypeRegistrationTimeout = 5 * time.Minute
 )
 
@@ -277,50 +115,4 @@ func WaitTypeRegistrationProgressStatusComplete(ctx context.Context, conn *cloud
 	}
 
 	return nil, err
-}
-
-func getDeletionReasons(ctx context.Context, conn *cloudformation.CloudFormation, stackID, requestToken string) ([]string, error) {
-	var failures []string
-
-	err := listStackEventsForOperation(ctx, conn, stackID, requestToken, func(e *cloudformation.StackEvent) {
-		if isFailedEvent(e) || isStackDeletionEvent(e) {
-			failures = append(failures, aws.StringValue(e.ResourceStatusReason))
-		}
-	})
-	return failures, err
-}
-
-func getRollbackReasons(ctx context.Context, conn *cloudformation.CloudFormation, stackID, requestToken string) ([]string, error) {
-	var failures []string
-	err := listStackEventsForOperation(ctx, conn, stackID, requestToken, func(e *cloudformation.StackEvent) {
-		if isFailedEvent(e) || isRollbackEvent(e) {
-			failures = append(failures, aws.StringValue(e.ResourceStatusReason))
-		}
-	})
-	return failures, err
-}
-
-func getFailures(ctx context.Context, conn *cloudformation.CloudFormation, stackID, requestToken string) ([]string, error) {
-	var failures []string
-
-	err := listStackEventsForOperation(ctx, conn, stackID, requestToken, func(e *cloudformation.StackEvent) {
-		if isFailedEvent(e) {
-			failures = append(failures, aws.StringValue(e.ResourceStatusReason))
-		}
-	})
-	return failures, err
-}
-
-func isFailedEvent(event *cloudformation.StackEvent) bool {
-	return strings.HasSuffix(aws.StringValue(event.ResourceStatus), "_FAILED") && event.ResourceStatusReason != nil
-}
-
-func isRollbackEvent(event *cloudformation.StackEvent) bool {
-	return strings.HasPrefix(aws.StringValue(event.ResourceStatus), "ROLLBACK_") && event.ResourceStatusReason != nil
-}
-
-func isStackDeletionEvent(event *cloudformation.StackEvent) bool {
-	return aws.StringValue(event.ResourceStatus) == cloudformation.ResourceStatusDeleteInProgress &&
-		aws.StringValue(event.ResourceType) == "AWS::CloudFormation::Stack" &&
-		event.ResourceStatusReason != nil
 }
